+++ test/integration/TestVerifiers.js
@@ -165,7 +165,7 @@ contract('Chain: 1 or more verifiers scenario (base on configuration)', (account
           it('should NOT be possible to reveal using wrong secret', async () => {
             const awaits = [];
             for (let i = 0; i < verifiersCount; i += 1) {
-              awaits.push(ministroChain.reveal(proposals[i], `${secrets[i]}_`, { from: verifiersAddr[i] }, true));
+              awaits.push(ministroChain.reveal(proposals[i], `0x${secrets[i].substring(2).split('').reverse().join('')}`, { from: verifiersAddr[i] }, true));
             }
             await Promise.all(awaits);
           });