+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+## [0.4.0] - 2019-03-14
+### Added
 - option to update `minimumStakingTokenPercentage`.
 - `initialBlockHeights` to set the initial block-height per shard
 - fetch VerifierRegistry contract from ContractRegistry